move common test code to test_utils
