+++ tests/unit/test_facts.py
@@ -3,31 +3,13 @@ from unittest.mock import patch
 
 import pytest
 import yaml
-from pandas import DataFrame
 from pybatfish.client.session import Session
-from pybatfish.datamodel.answer import TableAnswer
-from pybatfish.question.question import QuestionBase
 
 from module_utils.bf_util import (
     get_facts,
     _encapsulate_nodes_facts, get_node_count, load_facts, validate_facts,
     write_facts, _unencapsulate_facts, assert_dict_subset)
-
-
-class MockTableAnswer(TableAnswer):
-    def __init__(self, frame_to_use=DataFrame()):
-        self._frame = frame_to_use
-
-    def frame(self):
-        return self._frame
-
-
-class MockQuestion(QuestionBase):
-    def __init__(self, answer=None):
-        self._answer = answer if answer is not None else MockTableAnswer()
-
-    def answer(self, *args, **kwargs):
-        return self._answer
+from tests.unit.test_utils import MockTableAnswer, MockQuestion
 
 
 def test_get_facts_questions():
